+++ apps/src/templates/sectionAssessments/StudentsAssessmentsMultipleChoiceTable.story.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import StudentMultipleChoiceOverviewTable from './StudentMultipleChoiceOverviewTable';
+import StudentsAssessmentsMultipleChoiceTable from './StudentsAssessmentsMultipleChoiceTable';
 import {studentOverviewData} from './assessmentsTestHelpers';
 
 export default storybook => {
   return storybook
-    .storiesOf('SectionAssessments/StudentMultipleChoiceOverviewTable', module)
+    .storiesOf('SectionAssessments/StudentsAssessmentsMultipleChoiceTable', module)
     .addStoryTable([
       {
         name: 'Student overview',
@@ -12,7 +12,7 @@ export default storybook => {
             The table shows how many questions each student answered correctly and a
             time stamp for when an assessment is submitted.`,
         story: () => (
-          <StudentMultipleChoiceOverviewTable
+          <StudentsAssessmentsMultipleChoiceTable
             studentOverviewData={studentOverviewData}
           />
         )
